modify code based on reviews

+++ src/services/user.service.js
@@ -3,6 +3,7 @@ import config from "../config/index.js";
 import APIError from "../helper/apiError.js";
 import httpStatus from "http-status";
 import sequelize from "../models/index.js";
+import { ApiDataResponse, ApiPaginatedResponse } from "../helper/apiResponse.js";
 
 const { User, Customer } = sequelize.models;
 
@@ -16,7 +17,10 @@ const login = async (payload) => {
         throw new APIError({ message: "Incorrect password !", status: httpStatus.UNAUTHORIZED });
     }
 
-    return jwt.sign({ id: user.id }, config.token_secret, { expiresIn: config.token_expiry });
+    const jwtToken = jwt.sign({ id: user.id }, config.token_secret, {
+        expiresIn: config.token_expiry,
+    });
+    return new ApiDataResponse(httpStatus.OK, "login success", { token: jwtToken });
 };
 
 const addUser = async (payload) => {
@@ -26,7 +30,9 @@ const addUser = async (payload) => {
     }
 
     const newUser = await User.create(payload);
-    return newUser.id;
+    delete newUser.dataValues.password;
+
+    return new ApiDataResponse(httpStatus.CREATED, "create success", newUser);
 };
 
 const getUserDetail = async (userId) => {
@@ -47,11 +53,13 @@ const getListUsers = async (pageIndex, pageSize) => {
     const users = await User.findAll({
         attributes: { exclude: ["password"] },
     });
-    if (!users.length) {
+
+    const totalCount = users.length;
+    if (!totalCount) {
         throw new APIError({ message: "Users not found !", status: httpStatus.NOT_FOUND });
     }
 
-    const totalPages = Math.ceil(users.length / pageSize);
+    const totalPages = Math.ceil(totalCount / pageSize);
     if (pageIndex > totalPages) {
         throw new APIError({ message: "Invalid page index", status: httpStatus.BAD_REQUEST });
     }
@@ -59,7 +67,13 @@ const getListUsers = async (pageIndex, pageSize) => {
     const startIndex = (pageIndex - 1) * pageSize;
     const endIndex = startIndex + pageSize;
 
-    return users.slice(startIndex, endIndex);
+    return new ApiPaginatedResponse(
+        pageIndex,
+        pageSize,
+        totalCount,
+        totalPages,
+        users.slice(startIndex, endIndex)
+    );
 };
 
 const updateUser = async (userId, payload) => {
@@ -68,7 +82,8 @@ const updateUser = async (userId, payload) => {
         throw new APIError({ message: "User not found", status: httpStatus.NOT_FOUND });
     }
 
-    return updatedUser.id;
+    delete updatedUser.dataValues.password;
+    return new ApiDataResponse(httpStatus.OK, "update success", updatedUser);
 };
 
 const activeUser = async (userId) => {
@@ -77,7 +92,8 @@ const activeUser = async (userId) => {
         throw new APIError({ message: "User not found", status: httpStatus.NOT_FOUND });
     }
 
-    return activatedUser.id;
+    delete activatedUser.dataValues.password;
+    return new ApiDataResponse(httpStatus.OK, "active success", activatedUser);
 };
 
 const inactiveUser = async (userId) => {
@@ -86,17 +102,17 @@ const inactiveUser = async (userId) => {
         throw new APIError({ message: "User not found", status: httpStatus.NOT_FOUND });
     }
 
-    return inactivatedUser.id;
+    delete inactivatedUser.dataValues.password;
+    return new ApiDataResponse(httpStatus.OK, "inactive success", inactivatedUser);
 };
 
 const hardDeleteUser = async (userId) => {
-    await User.destroy({
-        where: {
-            id: userId,
-        },
-    });
+    const deletedUser = await User.destroy({ where: { id: userId } });
+    if (!deletedUser) {
+        throw new APIError({ message: "User not found", status: httpStatus.NOT_FOUND });
+    }
 
-    return userId;
+    return new ApiDataResponse(httpStatus.OK, "hard delete success", deletedUser);
 };
 
 export {
